R Markdown file structured like gemma2
This markdown file is how I initially used gemma2 as a template to add my individual variance option. This modification also extends beyond gemma2 in that I attempted to translate the functions MphCalcP and MphCalcBeta from C++ to R, but not as successfully due to my confusion about the quantity W and c_size in those functions.